Fix merger rules of policies group issue

+++ examples/ex03_value_policy.cpp
@@ -29,16 +29,16 @@ int main() {
                                  policy::error::expected>;
 
   // Same overflow input for all three policies.
-  auto const lhs_checked = checked_t{65530};
-  auto const rhs_checked = checked_t{20};
+  auto const lhs_checked = checked_t{static_cast<std::uint16_t>(65530)};
+  auto const rhs_checked = checked_t{static_cast<std::uint16_t>(20)};
   auto const checked_result = operations::add(lhs_checked, rhs_checked);
 
-  auto const lhs_unchecked = unchecked_t{65530};
-  auto const rhs_unchecked = unchecked_t{20};
+  auto const lhs_unchecked = unchecked_t{static_cast<std::uint16_t>(65530)};
+  auto const rhs_unchecked = unchecked_t{static_cast<std::uint16_t>(20)};
   auto const unchecked_result = operations::add(lhs_unchecked, rhs_unchecked);
 
-  auto const lhs_saturating = saturating_t{65530};
-  auto const rhs_saturating = saturating_t{20};
+  auto const lhs_saturating = saturating_t{static_cast<std::uint16_t>(65530)};
+  auto const rhs_saturating = saturating_t{static_cast<std::uint16_t>(20)};
   auto const saturating_result =
       operations::add(lhs_saturating, rhs_saturating);
 
@@ -48,11 +48,13 @@ int main() {
     std::cerr << "checked policy should report overflow\n";
     return 1;
   }
-  if (!unchecked_result.has_value() || unchecked_result->value() != 14) {
+  if (!unchecked_result.has_value() ||
+      unchecked_result->value() != static_cast<std::uint16_t>(14)) {
     std::cerr << "unchecked policy should wrap\n";
     return 1;
   }
-  if (!saturating_result.has_value() || saturating_result->value() != 65535) {
+  if (!saturating_result.has_value() ||
+      saturating_result->value() != static_cast<std::uint16_t>(65535)) {
     std::cerr << "saturating policy should clamp\n";
     return 1;
   }
